Move BaseServiceClient, BaseServiceClient<T>, ISupportsListing and QueryValues to Core namespace/folder
