+++ src/FaluSdk/Core/BasicListOptions.cs
@@ -1,5 +1,4 @@
-﻿using Falu.Infrastructure;
-using System;
+﻿using System;
 
 namespace Falu.Core
 {
