+++ src/FaluSdk/FileUploads/FileUploadsServiceClient.cs
@@ -1,5 +1,4 @@
 ﻿using Falu.Core;
-using Falu.Infrastructure;
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
