+++ src/FaluSdk/MoneyBalances/MoneyBalancesServiceClient.cs
@@ -1,5 +1,4 @@
 ﻿using Falu.Core;
-using Falu.Infrastructure;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
