+++ src/FaluSdk/PaymentRefunds/PaymentRefundsServiceClient.cs
@@ -1,5 +1,4 @@
 ﻿using Falu.Core;
-using Falu.Infrastructure;
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading;
